requirements.txt
langchain
openai
faiss-cpu
streamlit
pypdf
python-dotenv
